+++ Portfolio.API/Application/DTOs/TransactionDto.cs
@@ -7,7 +7,7 @@ public class TransactionDto : BaseDto
 {
     public WalletDto Wallet { get; set; } = new WalletDto();
     public string Symbol { get; set; } = string.Empty;
-    public double Amount { get; set; }
-    public double PriceAtTransaction { get; set; }
+    public decimal Amount { get; set; }
+    public decimal PriceAtTransaction { get; set; }
     public TransactionType TransactionType { get; set; }
 }
