+++ Portfolio.API/Application/DTOs/TransferAssetDto.cs
@@ -5,5 +5,5 @@ public class TransferAssetDto
     public Guid FromWalletId { get; set; }
     public string TargetWalletAddress { get; set; } = string.Empty;
     public string Symbol { get; set; } = string.Empty;
-    public double AssetAmount { get; set; }
+    public decimal AssetAmount { get; set; }
 }
